+++ CHANGELOG.md
@@ -6,7 +6,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ## [ForMIDI development version]
 
 ### Added
-
+- `build.sh` can now also use the Intel ifx compiler: `$ FC='ifx' ./build.sh`. the default compiler is GFortran.
 
 ## [ForMIDI 0.1 "Formidable"] 2021-03-02